change some integer fields to float

+++ Backend/models/transaction_model.py
@@ -6,7 +6,7 @@ from django.utils import timezone
 
 class Transactions(models.Model):
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
-    total_revenue=models.IntegerField(validators=[MaxValueValidator(100000)])
+    total_revenue=models.FloatField(validators=[MaxValueValidator(100000)])
     total_sold_items=models.IntegerField(validators=[MaxValueValidator(10000)])
     checkoutDate=models.DateTimeField(default=timezone.now)
     user=models.ForeignKey('custom_user.user',on_delete=models.CASCADE)
